Sprint 22 (add playlist to db)

+++ app/src/main/java/com/practicum/playlistmaker_1/media_library/domain/impl/FavouriteTracksInteractorImpl.kt
@@ -13,7 +13,7 @@ class FavouriteTracksInteractorImpl(
         return favouriteTracksRepository.getFavoritesTracks()
     }
 
-    override suspend fun isFavoriteTrack(trackId: Int): Flow<Boolean> {
+    override suspend fun isFavoriteTrack(trackId: Long): Flow<Boolean> {
         return favouriteTracksRepository.isFavoriteTrack(trackId)
     }
 
@@ -21,7 +21,7 @@ class FavouriteTracksInteractorImpl(
         favouriteTracksRepository.addToFavorites(track)
     }
 
-    override suspend fun deleteFromFavorites(trackId: Int) {
+    override suspend fun deleteFromFavorites(trackId: Long) {
         favouriteTracksRepository.deleteFromFavorites(trackId)
     }
 }
